requirements.txt
fastapi
uvicorn
supabase
pandas
numpy
scikit-learn
prophet
python-dotenv
pydantic
